// src/types/frontend/guarantor.ts
export default interface Guarantor {
  guarantorAddress: string;
  guarantorBirthDate: string;
  guarantorDocumentNumber: string;
  guarantorDocumentType: string;
  guarantorEmail: string;
  guarantorFirstName: string;
  guarantorLastName: string;
  guarantorPhone: string;
  guarantorPopulation: string;
  guarantorPostalCode: string;
  guarantorProvince: string;
}